backend/pyproject.toml
[tool.poetry]
name = "islah"
version = "0.1.0"
description = ""
authors = ["ayouba <ayoubadoumbia041@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = ">=0.115.8,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
pydantic = { version = ">=2.10.6,<3.0.0", extras = ["email"] }
httpx = ">=0.28.1,<0.29.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.38"
psycopg2-binary = "^2.9.10"
pyyaml = "^6.0.2"
websockets = "^15.0"
watchfiles = "^1.0.4"
alembic = "^1.14.1"
pandas = "^2.2.3"
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
pydantic-settings = "^2.8.1"


[tool.poetry.scripts]
setup = "src.setup_docker:setup"
stop = "src.setup_docker:stop_docker"
reset = "src.setup_docker:reset_docker"
docker = "src.setup_docker:build_and_run_docker"
startapp = "src.main:main"
logs = "src.setup_docker:show_docker_logs"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
